backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.43
alembic==1.12.1
psycopg2-binary==2.9.10
redis==5.0.1
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-cognitiveservices-language-textanalytics==5.3.0
openai==1.3.8
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
aiofiles==23.2.1
pillow==10.1.0
pypdf2==3.0.1
python-docx==1.1.0
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pytesseract==0.3.10
easyocr==1.7.0
pdf2image==1.17.0
PyMuPDF==1.23.0
# Marine Weather and Routing Enhancements
geopy==2.4.1
shapely==2.0.2
folium==0.15.1
meteomatics==1.0.0
stormglass==0.1.0
noaa-coops==0.1.0
networkx==3.2.1
scipy==1.11.4
